Separate legacy import conflict policy
Move exact-row matching and recoverable content reconciliation into a focused importer module while preserving SQL, outcomes, and the public import API.
